[EXT][APP]:1、修复存储芯片外设读的参数

+++ ext-driver/Ext_MB85RS64.h
@@ -17,6 +17,6 @@
 
 extern void FRAM_Erase(uint16 EraseAddr,uint16 len);
 extern void FRAM_Write(uint8 *tx_buffer,uint16 WriteAddr,uint16 len);
-extern void FRAM_Read(uint8 *rx_buffer,uint8 ReadAddr,uint16 len);
+extern void FRAM_Read(uint8 *rx_buffer,uint16 ReadAddr,uint16 len);
 
 #endif
